started to refactor predicates out of SearchBuilder
