+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/predicate/SearchBuilderJoinEnum.java
@@ -1,0 +1,13 @@
+package ca.uhn.fhir.jpa.dao.predicate;
+
+public enum SearchBuilderJoinEnum {
+	DATE,
+	NUMBER,
+	QUANTITY,
+	REFERENCE,
+	STRING,
+	TOKEN,
+	URI,
+	COORDS
+
+}
