+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/predicate/SearchBuilderTokenModeEnum.java
@@ -1,0 +1,7 @@
+package ca.uhn.fhir.jpa.dao.predicate;
+
+public enum SearchBuilderTokenModeEnum {
+	SYSTEM_ONLY,
+	VALUE_ONLY,
+	SYSTEM_AND_VALUE
+}
